Update update.py
This is a py file used to collected replies for a post, the replies of that reply and match them in a version of post id, reply userid, reply top 20 words, user id replies of that reply, sub-reply top 20 words. Use top 20 words to as an id for the reply because the reply id in facebook is dynamic.